// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type AbilityScore = {
	__typename?: 'AbilityScore';
	_id: Scalars['String'];
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	full_name?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	skills: Array<Skill>;
	url?: Maybe<Scalars['String']>;
};

export type AbilityScoreSkillsArgs = {
	limit?: InputMaybe<Scalars['Int']>;
	skip?: InputMaybe<Scalars['Int']>;
	sort?: InputMaybe<SortFindManySkillInput>;
};

export type AbilityScoreSkills = {
	__typename?: 'AbilityScoreSkills';
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type AbilityScoreSkillsInput = {
	index?: InputMaybe<Scalars['String']>;
	name?: InputMaybe<Scalars['String']>;
	url?: InputMaybe<Scalars['String']>;
};

export type Alignment = {
	__typename?: 'Alignment';
	_id: Scalars['String'];
	abbreviation?: Maybe<Scalars['String']>;
	desc?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type Background = {
	__typename?: 'Background';
	_id: Scalars['String'];
	bonds?: Maybe<BackgroundPersonality_Traits>;
	feature?: Maybe<BackgroundFeature>;
	flaws?: Maybe<BackgroundPersonality_Traits>;
	ideals?: Maybe<BackgroundIdeals>;
	index?: Maybe<Scalars['String']>;
	language_options?: Maybe<BackgroundLanguage_Options>;
	name?: Maybe<Scalars['String']>;
	personality_traits?: Maybe<BackgroundPersonality_Traits>;
	starting_equipment?: Maybe<Array<Maybe<BackgroundStarting_Equipment>>>;
	starting_equipment_options?: Maybe<Array<Maybe<BackgroundStarting_Equipment_Options>>>;
	starting_proficiencies: Array<Proficiency>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type Class = {
	__typename?: 'Class';
	_id: Scalars['String'];
	class_levels: Array<Level>;
	hit_die?: Maybe<Scalars['Float']>;
	index?: Maybe<Scalars['String']>;
	multi_classing?: Maybe<ClassMulti_Classing>;
	name?: Maybe<Scalars['String']>;
	proficiencies: Array<Proficiency>;
	proficiency_choices?: Maybe<Array<Maybe<ClassMulti_ClassingProficiency_Choices>>>;
	saving_throws?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	spellcasting?: Maybe<ClassSpellcasting>;
	spells: Array<Spell>;
	starting_equipment?: Maybe<Array<Maybe<ClassStarting_Equipment>>>;
	starting_equipment_options?: Maybe<Array<Maybe<ClassStarting_Equipment_Options>>>;
	subclasses?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type Condition = {
	__typename?: 'Condition';
	_id: Scalars['String'];
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type DamageType = {
	__typename?: 'DamageType';
	_id: Scalars['String'];
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type Equipment = {
	__typename?: 'Equipment';
	_id: Scalars['String'];
	armor_category?: Maybe<Scalars['String']>;
	armor_class?: Maybe<EquipmentArmor_Class>;
	capacity?: Maybe<Scalars['String']>;
	category_range?: Maybe<Scalars['String']>;
	contents?: Maybe<Array<Maybe<EquipmentContents>>>;
	cost?: Maybe<EquipmentCost>;
	damage?: Maybe<EquipmentDamage>;
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	equipment_category?: Maybe<AbilityScoreSkills>;
	gear_category?: Maybe<AbilityScoreSkills>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	properties: Array<WeaponProperty>;
	quantity?: Maybe<Scalars['Float']>;
	range?: Maybe<EquipmentRange>;
	special?: Maybe<Array<Maybe<Scalars['String']>>>;
	speed?: Maybe<EquipmentSpeed>;
	stealth_disadvantage?: Maybe<Scalars['Boolean']>;
	str_minimum?: Maybe<Scalars['Float']>;
	throw_range?: Maybe<EquipmentThrow_Range>;
	tool_category?: Maybe<Scalars['String']>;
	two_handed_damage?: Maybe<EquipmentTwo_Handed_Damage>;
	url?: Maybe<Scalars['String']>;
	vehicle_category?: Maybe<Scalars['String']>;
	weapon_category?: Maybe<Scalars['String']>;
	weapon_range?: Maybe<Scalars['String']>;
	weight?: Maybe<Scalars['Float']>;
};

export type EquipmentPropertiesArgs = {
	limit?: InputMaybe<Scalars['Int']>;
	skip?: InputMaybe<Scalars['Int']>;
	sort?: InputMaybe<SortFindManyWeaponPropertyInput>;
};

export type EquipmentArmor_Class = {
	__typename?: 'EquipmentArmor_class';
	base?: Maybe<Scalars['Float']>;
	dex_bonus?: Maybe<Scalars['Boolean']>;
	max_bonus?: Maybe<Scalars['Float']>;
};

export type EquipmentCategory = {
	__typename?: 'EquipmentCategory';
	_id: Scalars['String'];
	equipment: Array<Equipment>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type Feat = {
	__typename?: 'Feat';
	_id: Scalars['String'];
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	prerequisites?: Maybe<Array<Maybe<FeatPrerequisites>>>;
	url?: Maybe<Scalars['String']>;
};

export type FeatPrerequisites = {
	__typename?: 'FeatPrerequisites';
	ability_score?: Maybe<AbilityScoreSkills>;
	minimum_score?: Maybe<Scalars['Float']>;
};

export type Feature = {
	__typename?: 'Feature';
	_id: Scalars['String'];
	choice?: Maybe<FeatureChoice>;
	class?: Maybe<Class>;
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	feature_specific?: Maybe<FeatureFeature_Specific>;
	index?: Maybe<Scalars['String']>;
	level?: Maybe<Scalars['Float']>;
	name?: Maybe<Scalars['String']>;
	parent?: Maybe<AbilityScoreSkills>;
	prerequisites?: Maybe<Array<Maybe<FeaturePrerequisites>>>;
	reference?: Maybe<Scalars['String']>;
	subclass?: Maybe<Subclass>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type Language = {
	__typename?: 'Language';
	_id: Scalars['String'];
	desc?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	script?: Maybe<Scalars['String']>;
	type?: Maybe<Scalars['String']>;
	typical_speakers?: Maybe<Array<Maybe<Scalars['String']>>>;
	url?: Maybe<Scalars['String']>;
};

export type Level = {
	__typename?: 'Level';
	_id: Scalars['String'];
	ability_score_bonuses?: Maybe<Scalars['Float']>;
	class?: Maybe<AbilityScoreSkills>;
	class_specific?: Maybe<LevelClass_Specific>;
	features: Array<Feature>;
	index?: Maybe<Scalars['String']>;
	level?: Maybe<Scalars['Float']>;
	prof_bonus?: Maybe<Scalars['Float']>;
	spellcasting?: Maybe<LevelSpellcasting>;
	subclass?: Maybe<AbilityScoreSkills>;
	subclass_specific?: Maybe<LevelSubclass_Specific>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type MagicItem = {
	__typename?: 'MagicItem';
	_id: Scalars['String'];
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	equipment_category?: Maybe<AbilityScoreSkills>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	rarity?: Maybe<MagicItemRarity>;
	url?: Maybe<Scalars['String']>;
	variant?: Maybe<Scalars['Boolean']>;
	variants?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
};

export type MagicItemRarity = {
	__typename?: 'MagicItemRarity';
	name?: Maybe<Scalars['String']>;
};

export type MagicSchool = {
	__typename?: 'MagicSchool';
	_id: Scalars['String'];
	desc?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type Monster = {
	__typename?: 'Monster';
	_id: Scalars['String'];
	actions?: Maybe<Array<Maybe<MonsterActions>>>;
	alignment?: Maybe<Scalars['String']>;
	armor_class?: Maybe<Scalars['Float']>;
	challenge_rating?: Maybe<Scalars['Float']>;
	charisma?: Maybe<Scalars['Float']>;
	condition_immunities?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	constitution?: Maybe<Scalars['Float']>;
	damage_immunities?: Maybe<Array<Maybe<Scalars['String']>>>;
	damage_resistances?: Maybe<Array<Maybe<Scalars['String']>>>;
	damage_vulnerabilities?: Maybe<Array<Maybe<Scalars['String']>>>;
	dexterity?: Maybe<Scalars['Float']>;
	forms?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	hit_dice?: Maybe<Scalars['String']>;
	hit_points?: Maybe<Scalars['Float']>;
	index?: Maybe<Scalars['String']>;
	intelligence?: Maybe<Scalars['Float']>;
	languages?: Maybe<Scalars['String']>;
	legendary_actions?: Maybe<Array<Maybe<MonsterLegendary_Actions>>>;
	name?: Maybe<Scalars['String']>;
	proficiencies?: Maybe<Array<Maybe<MonsterProficiencies>>>;
	reactions?: Maybe<Array<Maybe<MonsterReactions>>>;
	senses?: Maybe<MonsterSenses>;
	size?: Maybe<Scalars['String']>;
	special_abilities?: Maybe<Array<Maybe<MonsterSpecial_Abilities>>>;
	speed?: Maybe<MonsterSpeed>;
	strength?: Maybe<Scalars['Float']>;
	subtype?: Maybe<Scalars['String']>;
	type?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
	wisdom?: Maybe<Scalars['Float']>;
	xp?: Maybe<Scalars['Float']>;
};
// ...
export type Proficiency = {
	__typename?: 'Proficiency';
	_id: Scalars['String'];
	classes?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	races?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	reference?: Maybe<ProficiencyReference>;
	type?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type ProficiencyReference = {
	__typename?: 'ProficiencyReference';
	index?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	type?: Maybe<Scalars['String']>;
	url?: Maybe<Scalars['String']>;
};

export type Query = {
	__typename?: 'Query';
	abilityScore?: Maybe<AbilityScore>;
	abilityScores: Array<AbilityScore>;
	alignment?: Maybe<Alignment>;
	alignments: Array<Alignment>;
	background?: Maybe<Background>;
	backgrounds: Array<Background>;
	class?: Maybe<Class>;
	classes: Array<Class>;
	condition?: Maybe<Condition>;
	conditions: Array<Condition>;
	damageType?: Maybe<DamageType>;
	damageTypes: Array<DamageType>;
	equipment?: Maybe<Equipment>;
	equipmentCategories: Array<EquipmentCategory>;
	equipmentCategory?: Maybe<EquipmentCategory>;
	equipments: Array<Equipment>;
	feat?: Maybe<Feat>;
	feats: Array<Feat>;
	feature?: Maybe<Feature>;
	features: Array<Feature>;
	language?: Maybe<Language>;
	languages: Array<Language>;
	level?: Maybe<Level>;
	levels: Array<Level>;
	magicItem?: Maybe<MagicItem>;
	magicItems: Array<MagicItem>;
	magicSchool?: Maybe<MagicSchool>;
	magicSchools: Array<MagicSchool>;
	monster?: Maybe<Monster>;
	monsters: Array<Monster>;
	proficiencies: Array<Proficiency>;
	proficiency?: Maybe<Proficiency>;
	race?: Maybe<Race>;
	races: Array<Race>;
	rule?: Maybe<Rule>;
	ruleSection?: Maybe<RuleSection>;
	ruleSections: Array<RuleSection>;
	rules: Array<Rule>;
	skill?: Maybe<Skill>;
	skills: Array<Skill>;
	spell?: Maybe<Spell>;
	spells: Array<Spell>;
	subclass?: Maybe<Subclass>;
	subclasses: Array<Subclass>;
	subrace?: Maybe<Subrace>;
	subraces: Array<Subrace>;
	trait?: Maybe<Trait>;
	traits: Array<Trait>;
	weaponProperties: Array<WeaponProperty>;
	weaponProperty?: Maybe<WeaponProperty>;
};
// ...
export type Race = {
	__typename?: 'Race';
	_id: Scalars['String'];
	ability_bonus_options?: Maybe<RaceAbility_Bonus_Options>;
	ability_bonuses?: Maybe<Array<Maybe<RaceAbility_Bonuses>>>;
	age?: Maybe<Scalars['String']>;
	alignment?: Maybe<Scalars['String']>;
	index?: Maybe<Scalars['String']>;
	language_desc?: Maybe<Scalars['String']>;
	language_options?: Maybe<RaceLanguage_Options>;
	languages?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	name?: Maybe<Scalars['String']>;
	size?: Maybe<Scalars['String']>;
	size_description?: Maybe<Scalars['String']>;
	speed?: Maybe<Scalars['Float']>;
	starting_proficiencies?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	starting_proficiency_options?: Maybe<RaceStarting_Proficiency_Options>;
	subraces?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	traits: Array<Trait>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type Spell = {
	__typename?: 'Spell';
	_id: Spell_Id;
	area_of_effect?: Maybe<SpellArea_Of_Effect>;
	attack_type?: Maybe<Scalars['String']>;
	casting_time?: Maybe<Scalars['String']>;
	classes?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	components?: Maybe<Array<Maybe<Scalars['String']>>>;
	concentration?: Maybe<Scalars['Boolean']>;
	damage?: Maybe<SpellDamage>;
	dc?: Maybe<SpellDc>;
	desc?: Maybe<Array<Maybe<Scalars['String']>>>;
	duration?: Maybe<Scalars['String']>;
	heal_at_slot_level?: Maybe<Scalars['JSON']>;
	higher_level?: Maybe<Array<Maybe<Scalars['String']>>>;
	index?: Maybe<Scalars['String']>;
	level: Scalars['Float'];
	material?: Maybe<Scalars['String']>;
	name?: Maybe<Scalars['String']>;
	range?: Maybe<Scalars['String']>;
	ritual?: Maybe<Scalars['Boolean']>;
	school?: Maybe<MagicSchool>;
	subclasses?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
	url?: Maybe<Scalars['String']>;
};
// ...
export type GetSpellsQuery = {
	__typename?: 'Query';
	spells: Array<{
		__typename?: 'Spell';
		index?: string | null;
		name?: string | null;
		components?: Array<string | null> | null;
		level: number;
		classes?: Array<{
			__typename?: 'AbilityScoreSkills';
			index?: string | null;
			name?: string | null;
		} | null> | null;
		subclasses?: Array<{
			__typename?: 'AbilityScoreSkills';
			index?: string | null;
			name?: string | null;
		} | null> | null;
	}>;
};